fix(replay): split streaming helpers to satisfy 800-line file limit
pkg/extproc/replay_tool_trace.go was 813 lines — 13 over the structural
lint limit of 800. Move the streaming tool-call fragment helpers
(replayStreaming*, mergeReplayStreaming*) into a new
replay_tool_trace_streaming.go in the same package. No behavior change.
